backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0